refactor: move some functions from parser.js to utils.js
